+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 * Feat: Read `uncaught.handler.enabled` property from the external configuration 
 * Feat: Resolve servername from the localhost address
 * Fix: Consider {{ auto }} as a default ip address (#1015) 
+* Fix: Set release and environment on Transactions (#1152) 
 
 # 4.0.0-alpha.2